+++ matchms/similarity/CosineGreedy.py
@@ -1,4 +1,3 @@
-from typing import Tuple
 import numpy as np
 from matchms.typing import SpectrumType
 from .BaseSimilarity import BaseSimilarity
@@ -68,7 +67,7 @@ class CosineGreedy(BaseSimilarity):
         self.mz_power = mz_power
         self.intensity_power = intensity_power
 
-    def pair(self, reference: SpectrumType, query: SpectrumType) -> Tuple[float, int]:
+    def pair(self, reference: SpectrumType, query: SpectrumType) -> np.ndarray:
         """Calculate cosine score between two spectra.
 
         Parameters