+++ matchms/similarity/IntersectMz.py
@@ -1,5 +1,5 @@
 import numpy as np
-from matchms.typing import SpectrumType
+from .. import Spectrum
 from .BaseSimilarity import BaseSimilarity
 
 
@@ -47,7 +47,7 @@ class IntersectMz(BaseSimilarity):
         """
         self.scaling = scaling
 
-    def pair(self, reference: SpectrumType, query: SpectrumType) -> float:
+    def pair(self, reference: Spectrum, query: Spectrum) -> np.ndarray:
         """This will calculate the similarity score between two spectra."""
         mz_ref = set(reference.peaks.mz)
         mz_query = set(query.peaks.mz)