+++ matchms/similarity/PrecursorMzMatch.py
@@ -1,10 +1,9 @@
 from typing import List
 import numpy as np
-from sparsestack import StackedSparseArray
 from matchms.similarity.spectrum_similarity_functions import (
     number_matching, number_matching_ppm, number_matching_symmetric,
     number_matching_symmetric_ppm)
-from matchms.typing import SpectrumType
+from matchms.Spectrum import Spectrum
 from .BaseSimilarity import BaseSimilarity
 
 
@@ -70,7 +69,7 @@ class PrecursorMzMatch(BaseSimilarity):
         assert tolerance_type in ["Dalton", "ppm"], "Expected type from ['Dalton', 'ppm']"
         self.type = tolerance_type
 
-    def pair(self, reference: SpectrumType, query: SpectrumType) -> float:
+    def pair(self, reference: Spectrum, query: Spectrum) -> np.ndarray:
         """Compare precursor m/z between reference and query spectrum.
 
         Parameters
@@ -91,8 +90,7 @@ class PrecursorMzMatch(BaseSimilarity):
         score = abs(precursormz_ref - precursormz_query)/mean_mz <= self.tolerance
         return np.asarray(score, dtype=self.score_datatype)
 
-    def matrix(self, references: List[SpectrumType], queries: List[SpectrumType],
-               array_type: str = "numpy",
+    def matrix(self, references: List[Spectrum], queries: List[Spectrum],
                is_symmetric: bool = False) -> np.ndarray:
         """Compare parent masses between all references and queries.
 
@@ -102,9 +100,6 @@ class PrecursorMzMatch(BaseSimilarity):
             List/array of reference spectra.
         queries
             List/array of Single query spectra.
-        array_type
-            Specify the output array type. Can be "numpy" or "sparse".
-            Default is "numpy" and will return a numpy array. "sparse" will return a COO-sparse array.
         is_symmetric
             Set to True when *references* and *queries* are identical (as for instance for an all-vs-all
             comparison). By using the fact that score[i,j] = score[j,i] the calculation will be about
@@ -133,12 +128,7 @@ class PrecursorMzMatch(BaseSimilarity):
         else:
             rows, cols, scores = number_matching_ppm(precursors_ref, precursors_query,
                                                      self.tolerance)
-        if array_type == "numpy":
-            scores_array = np.zeros((len(precursors_ref), len(precursors_query)))
-            scores_array[rows, cols] = scores.astype(self.score_datatype)
-            return scores_array
-        if array_type == "sparse":
-            scores_array = StackedSparseArray(len(precursors_ref), len(precursors_query))
-            scores_array.add_sparse_data(rows, cols, scores.astype(self.score_datatype), "")
-            return scores_array
-        return ValueError("`array_type` can only be 'numpy' or 'sparse'.")
+
+        scores_array = np.zeros((len(precursors_ref), len(precursors_query)))
+        scores_array[rows, cols] = scores.astype(self.score_datatype)
+        return scores_array